[query] remove non-primitives from scala fs
